MAINT: Move pytesttester outside of np.testing, to avoid creating unnecessary import dependencies
pytesttester is used by every single subpackage, so making it depend on np.testing just creates cyclic dependencies that can lead to circular imports

Relates to #11457
